[MDB IGNORE] [IDB IGNORE] WIP TGU (#1427)
Several months worth of updates.

---------

+++ code/datums/status_effects/debuffs/debuffs.dm
@@ -313,6 +313,12 @@
 	tick_interval = 0.4 SECONDS
 	alert_type = /atom/movable/screen/alert/status_effect/his_wrath
 
+/datum/status_effect/his_wrath/on_creation(mob/living/new_owner, His, Him)
+	. = ..()
+	linked_alert.name = "[His] Wrath"
+	linked_alert.desc = "You fled from [His] Grace instead of feeding [Him], and now you suffer."
+	linked_alert.icon_state = "[LOWER_TEXT(His)]_grace"
+
 /atom/movable/screen/alert/status_effect/his_wrath
 	name = "His Wrath"
 	desc = "You fled from His Grace instead of feeding Him, and now you suffer."
@@ -364,8 +370,8 @@
 		return FALSE
 
 	marked_underlay = new()
-	marked_underlay.pixel_x = -owner.pixel_x
-	marked_underlay.pixel_y = -owner.pixel_y
+	marked_underlay.pixel_w = -(owner.base_pixel_x + owner.base_pixel_w)
+	marked_underlay.pixel_z = -(owner.base_pixel_y + owner.base_pixel_z)
 	marked_underlay.transform *= 0.5
 	owner.vis_contents += marked_underlay
 	animate(marked_underlay, ready_delay, transform = matrix() * 1.2, flags = CIRCULAR_EASING | EASE_IN)
@@ -386,7 +392,7 @@
 	name = "Crusher mark underlay"
 	icon = 'icons/effects/effects.dmi'
 	icon_state = "shield"
-	appearance_flags = TILE_BOUND|LONG_GLIDE|RESET_COLOR|PIXEL_SCALE
+	appearance_flags = TILE_BOUND|LONG_GLIDE|RESET_COLOR|PIXEL_SCALE|KEEP_APART
 	vis_flags = VIS_UNDERLAY
 	mouse_opacity = MOUSE_OPACITY_TRANSPARENT
 
@@ -397,9 +403,7 @@
 	stack_threshold = 10
 	max_stacks = 10
 	overlay_file = 'icons/effects/bleed.dmi'
-	underlay_file = 'icons/effects/bleed.dmi'
 	overlay_state = "bleed"
-	underlay_state = "bleed"
 	var/bleed_damage = 200
 
 /datum/status_effect/stacking/saw_bleed/fadeout_effect()
@@ -441,7 +445,7 @@
 	for(var/datum/wound/bleeding_thing as anything in throat.wounds)
 		var/datum/wound_pregen_data/pregen_data = GLOB.all_wound_pregen_data[bleeding_thing.type]
 
-		if(pregen_data.wounding_types_valid(list(WOUND_SLASH)) && bleeding_thing.severity > WOUND_SEVERITY_MODERATE && bleeding_thing.blood_flow > 0)
+		if(pregen_data.wounding_types_valid(WOUND_SLASH) && bleeding_thing.severity > WOUND_SEVERITY_MODERATE && bleeding_thing.blood_flow > 0)
 			still_bleeding = TRUE
 			break
 	if(!still_bleeding)
@@ -725,22 +729,36 @@
 	. = ..()
 	direction = pick(NORTH, SOUTH, EAST, WEST)
 	new_owner.setDir(direction)
+	owner.add_traits(list(TRAIT_NO_FLOATING_ANIM, TRAIT_MOVE_FLYING), TRAIT_STATUS_EFFECT(id)) //I believe I can fly!
 
 /datum/status_effect/go_away/tick(seconds_between_ticks)
 	owner.AdjustStun(1, ignore_canstun = TRUE)
-	var/turf/T = get_step(owner, direction)
-	owner.forceMove(T)
+	var/turf/turf = get_step(owner, direction)
+	if(!turf)
+		qdel(src)
+		return
+	owner.forceMove(turf)
 
+/datum/status_effect/go_away/on_remove()
+	. = ..()
+	owner.remove_traits(list(TRAIT_NO_FLOATING_ANIM, TRAIT_MOVE_FLYING), TRAIT_STATUS_EFFECT(id))
+
+///Subtype of the go away effect (phases the mob in one direction) that deletes the owner on z-level change or when the time's up.
 /datum/status_effect/go_away/deletes_mob
 	id = "go_away_deletes_mob"
-	duration = INFINITY
+	duration = 30 SECONDS
 
-/datum/status_effect/go_away/deluxe/on_creation(mob/living/new_owner, set_duration)
+/datum/status_effect/go_away/deletes_mob/on_creation(mob/living/new_owner, set_duration)
 	. = ..()
 	RegisterSignal(new_owner, COMSIG_MOVABLE_Z_CHANGED, PROC_REF(wipe_bozo))
 
-/datum/status_effect/go_away/deluxe/proc/wipe_bozo()
-	qdel(owner)
+/datum/status_effect/go_away/deletes_mob/proc/wipe_bozo()
+	qdel(src)
+
+/datum/status_effect/go_away/deletes_mob/on_remove()
+	. = ..()
+	if(!QDELETED(owner))
+		qdel(owner)
 
 /atom/movable/screen/alert/status_effect/go_away
 	name = "TO THE STARS AND BEYOND!"
@@ -851,8 +869,11 @@
 
 /datum/status_effect/ants/proc/ants_washed()
 	SIGNAL_HANDLER
-	owner.remove_status_effect(/datum/status_effect/ants)
-	return COMPONENT_CLEANED
+
+	. = NONE
+
+	if(owner.remove_status_effect(/datum/status_effect/ants))
+		return COMPONENT_CLEANED|COMPONENT_CLEANED_GAIN_XP
 
 /datum/status_effect/ants/get_examine_text()
 	return span_warning("[owner.p_They()] [owner.p_are()] covered in ants!")
@@ -921,12 +942,14 @@
 	duration = 30 SECONDS
 	tick_interval = 1 SECONDS
 	alert_type = null
+	/// By how much we should increase the attack cooldown
+	var/cd_increase = 2.5
 
 /datum/status_effect/rebuked/on_apply()
 	owner.next_move_modifier *= 2
 	if(ishostile(owner))
 		var/mob/living/simple_animal/hostile/simple_owner = owner
-		simple_owner.ranged_cooldown_time *= 2.5
+		simple_owner.ranged_cooldown_time *= cd_increase
 	return TRUE
 
 /datum/status_effect/rebuked/on_remove()
@@ -936,7 +959,7 @@
 	owner.next_move_modifier *= 0.5
 	if(ishostile(owner))
 		var/mob/living/simple_animal/hostile/simple_owner = owner
-		simple_owner.ranged_cooldown_time /= 2.5
+		simple_owner.ranged_cooldown_time /= cd_increase
 
 /datum/status_effect/freezing_blast
 	id = "freezing_blast"
@@ -1075,6 +1098,38 @@
 	owner.remove_actionspeed_modifier(ACTIONSPEED_ID_MIDAS_BLIGHT, update = TRUE)
 	UnregisterSignal(owner, COMSIG_ATOM_UPDATE_OVERLAYS)
 	owner.update_icon()
+
+// Desginated Target - Applied typically by Flare lasers
+
+/atom/movable/screen/alert/status_effect/designated_target
+	name = "Designated Target"
+	desc = "You've been lit up by some kind of bright energy! Wash it off to get rid of it, or you'll be a lot easier to hit!"
+	icon_state = "designated_target"
+
+/datum/status_effect/designated_target
+	id = "designated_target"
+	duration = 2 MINUTES
+	alert_type = /atom/movable/screen/alert/status_effect/designated_target
+	status_type = STATUS_EFFECT_REFRESH
+	/// Dummy lighting object for our flare attached to our mob
+	var/obj/effect/dummy/lighting_obj/moblight/mob_flare
+
+/datum/status_effect/designated_target/on_apply()
+	mob_flare = owner.mob_light(3, 15, LIGHT_COLOR_FLARE)
+	ADD_TRAIT(owner, TRAIT_DESIGNATED_TARGET, id)
+	owner.add_filter("designated_target", 3, list("type" = "outline", "color" = COLOR_RED, "size" = 1))
+	return TRUE
+
+/datum/status_effect/designated_target/tick(seconds_between_ticks)
+	// If we are ever wet, remove our flare status effect
+	var/datum/status_effect/fire_handler/wet_stacks/splashed_with_water = locate() in owner.status_effects
+	if(istype(splashed_with_water))
+		qdel(src)
+
+/datum/status_effect/designated_target/on_remove()
+	QDEL_NULL(mob_flare)
+	owner.remove_filter("designated_target")
+	REMOVE_TRAIT(owner, TRAIT_DESIGNATED_TARGET, id)
 
 #undef HEALING_SLEEP_DEFAULT
 #undef HEALING_SLEEP_ORGAN_MULTIPLIER
